test: convert test_ber to TAP under t/
