+++ t/ber.t
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec "$(dirname "$0")/test_ber"
